+++ lib/custom_faqs.dart
@@ -1,3 +1,4 @@
-library custom_faqs;
-
+export 'package:custom_faqs/enums/custom_faq_style.dart';
+export 'package:custom_faqs/models/custom_faq_theme.dart';
+export 'package:custom_faqs/models/faq.dart';
 export 'package:custom_faqs/pages/custom_faq.page.dart';